Create Data.py
This program is simple, and it captures the basic details of the child. The create_plan function will have the logic to create a personalized SLP practice plan based on the child's interests, goals, and routine. More details can be added to the class and function depending on what kind of SLP practice regimen you want to create.

Remember to replace the 'example.com' links with actual useful links.